[Wayland] Anchor top edge of menu to center
Anchor only the top edge of the menu to the center when
--center mode is used instead of centering the entire menu. This
prevents the menu from bouncing around as the list is shortened.

For this to work we only anchor to the top instead of the top and
the bottom and set a top margin equal to half of the display width.
The max height of the menu also needs to be halved to prevent drawing
a menu that overflows the bottom edge of the screen.