Apply initial filter before evalutating accept-single
Follow up to 0a44fb65cdc052aeb53ee05ede05ac77ad2d2a0d
Which didn't actually result in the intended behavior of skipping all
menu interaction when `accept-single` is true and the initial filter
results in a single item.

This has the side effect that `ifne` now evaluates on the filtered items
as well.

Signed-off-by: Robert Günzler <r@gnzler.io>