Fix fallback to xclip
popen() returns NULL if the underlying fork() or pipe() fails, not if
the called program didn't run properly. The old code didn't actually
fall back to xclip on X11, it was never called.

To fix this, inspect the exit code of the programs that we run. We use
the output of the first program that exited cleanly.

CC #363.