Avoid glitch on first frame
With the recent wlroots scene-graph api changes, we no longer receive
the wl_display@enter output event before the very first frame. Meaning
that the first draw max_height is 640. This cause a visual glitch, and
a useless loop.

With this change, we use instead the minimum height of the outputs.